move cron into main module
